Feature/user

+++ src/main/java/com/Zlibrary/controller/user/UserController.java
@@ -10,6 +10,7 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.github.xiaoymin.knife4j.annotations.ApiSupport;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -27,11 +28,12 @@ public class UserController {
 
     private final UserService userService;
     private final UserMapper userMapper;
+    private final RedisTemplate<String, String> redisTemplate;
 
-
-    public UserController(UserService userService, UserMapper userMapper) {
+    public UserController(UserService userService, UserMapper userMapper, RedisTemplate<String, String> redisTemplate) {
         this.userService = userService;
         this.userMapper = userMapper;
+        this.redisTemplate = redisTemplate;
     }
 
     // 判断用户是否被删除
@@ -51,17 +53,31 @@ public class UserController {
     public ResultData SelByPage(@RequestParam(value = "page", required = false, defaultValue = "1") Integer currentPage,
                                 @RequestParam(value = "size", required = false, defaultValue = "10") Integer pageSize) {
 
+        // 判断参数是否为空
+        if (pageSize == null || currentPage == null) {
+            return ResultData.fail("页码为0");
+        }
+
+        // 校验分页参数是否合法
+        if (pageSize < 1 || pageSize > 100) {
+            return ResultData.fail("页码须在1到100之间");
+        }
+        if (currentPage < 1) {
+            return ResultData.fail("页码须大于1");
+        }
+
+        // 校验
         // 只查询未删除的用户和已激活的用户
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("delete_status", 0);
-        queryWrapper.eq("activation_status", 1);
+        queryWrapper.eq("delete_status", 0)
+                .eq("activation_status", 1);
 
         // 分页查询
         Page<User> page = new Page<>(currentPage, pageSize);
         // 条件queryWrapper
         Page<User> userPage = userMapper.selectPage(page, queryWrapper);
-        new PageBean<>(userPage.getTotal(), userPage.getRecords());
-        return ResultData.success(new PageBean<>(userPage.getTotal(), userPage.getRecords()));
+        PageBean<User> userPageBean = new PageBean<>(userPage.getTotal(), userPage.getRecords());
+        return ResultData.success(userPageBean);
 
     }
 
@@ -75,16 +91,17 @@ public class UserController {
     @GetMapping("/findById")
     public ResultData findById(@RequestParam(value = "userId") Integer id) {
 
-        // 判断用户是否存在，existsById方法封装在UserService中
-        if (!userService.existsById(id)) {
-            return ResultData.fail(404, "该用户不存在");
-        }
-        if (userService.getById(id).getDeleteStatus() == 1) {
-            return ResultData.fail(404, "该用户不存在");
-        } else {
-            return ResultData.success(userService.getById(id));
-        }
+        User userId = userService.getById(id);
 
+        // 判断用户是否存在，existsById方法封装在UserService中
+        if (userId == null) {
+            return ResultData.fail(404, "该用户不存在");
+        }
+        if (userId.getDeleteStatus() == 1) {
+            return ResultData.fail(404, "该用户不存在");
+        }
+        User byId = userService.getById(id);
+        return ResultData.success(byId);
 
     }
 
@@ -104,7 +121,7 @@ public class UserController {
             return ResultData.fail(500, "邮箱重复");
         } else {
             userService.save(user);
-            return ResultData.success(200, "add success");
+            return ResultData.success(200, "添加成功");
         }
 
     }
@@ -128,10 +145,10 @@ public class UserController {
 
         if (isUserDeleted(oldUser)) {
             userService.updateById(user);
-            return ResultData.success(200, "update success");
+            return ResultData.success(200, "修改失败");
         } else {
             if (!user.getEmail().equals(email)) {
-                return ResultData.fail(400, "Cannot update email.");
+                return ResultData.fail(400, "新旧邮箱不能一致");
             }
             return ResultData.fail(400, "该用户不存在");
         }
@@ -149,7 +166,9 @@ public class UserController {
     @PutMapping("/updateEmail")
     public ResultData updateEmail(@RequestParam Integer id,
                                   @RequestParam String oldEmail,
-                                  @RequestParam String newEmail) {
+                                  @RequestParam String newEmail,
+                                  @RequestParam ("email") String email,
+                                  @RequestParam ("code") String code) {
 
         // 获取用户
         User userId = userService.getById(id);
@@ -170,10 +189,21 @@ public class UserController {
         if (newUser != null) {
             return ResultData.fail(400, "新邮箱已被占用");
         }
-        // 更新用户信息，将旧邮箱替换成新邮箱
-        oldUser.setEmail(newEmail);
-        userService.updateById(oldUser);
-        return ResultData.success(200, "换绑邮箱成功");
+
+        // redis的key值
+        String prefix = "email_";
+        String key = prefix + email;
+        // 读取redis上的code值
+        String codeInRedis = redisTemplate.opsForValue().get(key);
+        if (codeInRedis != null && codeInRedis.equals(code)) {
+            // 验证码通过后删除key
+            redisTemplate.delete(key);
+            oldUser.setEmail(newEmail);
+            userService.updateById(oldUser);
+            return ResultData.success("换绑成功");
+        } else {
+            return ResultData.fail("验证错误");
+        }
     }
 
 
@@ -183,12 +213,16 @@ public class UserController {
      * @param id 用户id
      * @return ResultData code,msg
      */
-    @ApiOperation(value = "根据ID删除用户信息", httpMethod = "GET")
+    @ApiOperation(value = "根据ID删除用户信息", httpMethod = "DELETE")
     @DeleteMapping("/deleteById")
     public ResultData deleteById(Integer id) {
 
+        User byId = userService.getById(id);
+        if (isUserDeleted(byId)) {
+            return ResultData.fail(400, "删除失败");
+        }
         userService.removeById(id);
-        return ResultData.success(200, "delete success");
+        return ResultData.success(200, "删除成功");
 
     }
